requirements.txt
requests
Django==4.2.4
djangorestframework